requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
python-docx
pypdf2
langchain
langchain-google-genai
langchain-community
fpdf
gtts